Update 15sp7 autoyast profiles to open sshd service in mini installation
Update 15sp7 autoyast profiles to open sshd service in minimal
installation, based on bsc#1241954, update the mini autoyast profiles.
Add an autoyast profile for testsuite autoyast_create_hdd_gnome on
power, the testsuite is unsing a minimal installation profile, add
a new one for it which includes desktop appliactions module selected.
And correct a profile of create_hdd_textmode_powervm.xml which has a
duplicated addon of python3 module by changing one of them to
sle-module-systems-management which is a new default selected module
on SLES15SP7.