remove_kernel_packages: Fix transactional workaround
The transactional-update tool does not pass the zypper exit code through
so trying to remove packages which are not installed will result
in jb failure. However, remove_kernel_packages() callers rely
on the package list defined inside the function. Make sure that the devel
packages are included in the return list.