Allow ltss containers to be downloaded
Prior pulling a released ltss container, the test needs to login the container
engine to registry.suse.com. This can be done by using the already
registered credentials in `/etc/zypp/credentials.d/SCCcredentials` or by
simply using the LTSS regcode that we already have present in the tests
and should work also for non-suse hosts if required.

- ticket: https://progress.opensuse.org/issues/181652