Add mutex to signal when HA barriers are ready to use
Currently, HA modules rely on barriers to sync test steps among nodes,
they do so by calling `barrier_wait()` API without checking whether
the barriers have been created. This can lead to unexpected failures
when jobs other than the parent job run first and attempt to call
`barrier_wait()` on a barrier which has not been created. In those
cases, tests will fail on a call to `bmwqemu::mydie()` as part of the
`_try_lock()` function in `lockapi`.

To avoid such failures, this commit introduces a new mutex called
`ha_barriers_ready` which is created at the end of the `ha/barrier_init`
test module after all the necessary barriers have been created. The
mutex is used in the `ha/wait_barriers` module before the call to
`barrier_wait()` with a call to `mutex_wait()` which will block the
process until the mutex is created and unlocked.