Replace hardcoded 'eth0' with network_utils::iface() function
Some HA modules as well as one line in `lib/hacluster` default to `eth0`
when the tests do not get a network interface name set in the setting
**SUT_NETDEVICE**. This is good enough for SLES 12 & 15 where SUTs have
rules allowing for persistently named network interfaces, but breaks
tests in SLES 16 where this is more dynamic.

This commit replace the default `eth0` values in HA modules and
`lib/hacluster` for a call to `network_utils::iface()` which should
return the first interface in the SUT different than `lo`.

Related ticket: https://jira.suse.com/browse/TEAM-10345