Fix race condition in xfs fs creation in ensa2 test
As described in commit 6be0607, the serial terminal is too fast and by the time
it reaches mkfs.xfs, the partition table is not ready.

Because I could not find a way to force a sync, I added a script_retry that will
cut down on the amount of failures.

Related Ticket: TEAM-10321