Install libnsl-stub1 for ASE and NetWeaver installation
SLE 16.0 does not ship with a version of `libnsl.so.1` (deprecated in 15),
but several third party products cannot be installed if the library is
missing. As a workaround SLE 16.0 does ship with a `libnsl-stub1`. More
details at:

https://susedoc.github.io/release-notes/slesap-16.0/html/release-notes/index.html#jsc-DOCTEAM-1849

SAP's ASE and NetWeaver require `libnsl`, so this commit modifies the
installation test modules to also install `libnsl-stub1`. Workaround
code for https://bugzilla.suse.com/show_bug.cgi?id=1236235 has been
removed as the installation of `libnsl-stub1` makes this not necessary.

Also included in the commit is the removal of `saptune solution apply
SAP-ASE` on ASE installation module, as the solution has been removed
from `saptune` in 16. Recommendation is to install ASE with the default
`SAP_Base` solution.