NFS mm schedule: Run stress_ng on Tumbleweed
The NFS mm tests got enabled against SLE16 where the PackageHub isn't
available yet. The nfs.yaml schedule however schedule stress_ng test
module which is installing the package using the PackageHub. We need
to then make sure we do not run stress_ng on sle16. Since there is
no negation in the conditional scheduler and we run the nfs mm tests
only on SLE16 and Tumbleweed now, the commit is scheduling stress_ng
only for Tumblewwed. This is a temporary fix to ensure proper executions
on both SLE16 and Tumbleweed without compromising the test scope for
the Tumbleweed