Remove the HanaSR code about peering (#22211)
Remove most of the code to deal with network peering to the IBSm that is
creating or destroing the peering using cloud cli and perl. Now this
feature is implemented in qe-sap-deployment Terroform code.
Remove two test module and reference to it in the YAML_SCHEDULE, remove
higher library code.
Leave in place low lovel perl library code about the cloud cli.