Change fencing selection in qesap regression (#22332)
Change all the conf.yaml used for qesap regression tests.
Introduce a new terraform variable in Azure conf file to select if to
deploy or not the iscsi VM.
Improve the description of all the files, reflecting the deployed
scenario.
Remove terraform variables about the private IP ranges for files not
deploying any network peering.
Remove two not used files.