agama: Improve IPMI installation profile
This improvement to the existing IPMI profile is based on the PVM
profile with additional enhancements for bare metal. There should be
a bootloader timeout to ensure that the grub menu needle is matched.
We can't use an infinite timeout because some machines have console
input issues and the return key is not registered. A value of 45 seconds
is taken from the SLE 15 AY profile. We also need to set GRUB_TERMINAL
to console to ensure that the grub menu is displayed properly. We do
the same in SLE 15 AY profiles.