Fix use of mutes in netweaver_nfs_setup.pm
In the netweaver_nfs_setup module, we lock the support_server_ready mutex and we
never unlock it. Additionally, this mutex is already used by other modules most
notably lib/opensusebasetest.pm, which makes this a bigger problem as we are
unable to add a reboot in any module past the point we lock the mutex, without
introducing a dead lock.

To fix this, we use a different mutex that is exclusive to the
netweaver_nfs_setup module.

Related Ticket: TEAM-10431