Create module to configure SELinux enforcement
SLE4SAP 16 introduces SELinux, but the default configuration causes errors
during the installation of SAP software.

This creates a module that applies the instructions in the draft release notes
for SLE4SAP 16, when the SELINUX_POLICY variable is set in either ´permissive´
or ´enforcing´, and disables SELinux when it is set to ´disabled´ to allow for
debugging.

Related Ticket: TEAM-10414