Fix for sysm related failures in MU
Problem cause: SCC_ADDONS has 'sysm' but there was no mapping
for 'sysm' to the full module name. So by default the first one in the
list got selected and due to recent changes in SCC the first one moved
from "base" to "ha". But the HA module needs an own registration
code which caused the failures e.g.
https://openqa.suse.de/tests/18109900#step/add_additional_products/1