Disable grub timeout in SLES for SAP and SLES+HA qcow2
In SLES for SAP and SLES+HA tests in 12 and 15, the `create_hdd` jobs
relied on the `installation/disable_grub_timeout` test module to
generate qcow2 files with systems that had no timeout enabled in grub.
This was helpful in cluster tests to control via the
`boot/boot_to_desktop` test module when the SUT would boot, and could
lead to issues in some of the tests if a small timeout is set in grub
(like the default 8 seconds timeout).

This commit adds `bootloader` section into the SLES+HA and SLES for SAP
JSONnet files with the property `stopOnBootMenu` set to `true` in order
create systems for further tests with the grub timeout disabled.

Also adding the `boot/boot_to_desktop` test module in the `create_hdd`
schedule for SLES+HA, as it will be needle to boot the SUT with the grub
timeout disabled. Not needed in SLES for SAP schedules as they have
`installation/grub_test` included already.

Related Ticket: https://jira.suse.com/browse/TEAM-10438