Create a new dedicated Rear schedule for 16
Create a new backup schedule that can be used to test Rear feature on
16.
The new schedule is derived from the rear_backup.yaml in the same
folder.
Functional and important changes, that are 16 specific, are about using Agama.
This version of the schedule also move schedule specific settings to the
schedule file from the JobGroup file.