Add ISCSI_ENABLED terraform variable
Add a new terraform variable named iscsi_enabled. It is controlled by a
openQA variable using the same name. For the moment the value is derived
by FENCING_MECHANISM. This is an intermediate commit, in a next one it
will be possible to remove FENCING_MECHIANISN and all the terraform
variables controlled by that.