kselftests: Add timeout option
45s is a default value in the kselftests thus this is kept. However at
times, this might be not enough which might be also SUT-dependent. For
instance the cgroup tests for mem ctl fail if there is lower number of
CPUs and setting higher number of CPUs results in the cpu tests to fail
as the default 45s timeout isn't sufficient. This patch allows to
set --override-timeout value to the given tests