Fail test in case of AVCs
Check for AVCs and in case there any found, mark the test as failed if
`AVC_FAIL_ON_DENIALS` is set.
It is expected this test will run on most of the user space tests and
SELinux supported products, but for now run this code only in
consoletests and container tests.

I had to re-arrange `rootless_podman` as after failure consequent tests
were not recovering well from the openQA taken snapshot

- ticket: https://progress.opensuse.org/issues/182384