Check agama installation if agama installer banner not shown
After guest installation console is re-attached, agama installer
might not be shown. For agama installation, it is a bit different
because agama installer shell is not a fully-fledged environment
and guest needs to be rebooted after installation finishes. So it
is necesary to still check agama installation progression in the
subroutine check_guest_installation_result_via_ssh and add a new
subroutine is_guest_installation_done.