Update Availability Zone in case of change
In case of failure to create VM in defined AZ test logic tries to loop over
other AZ in same region and create VM in them. In case of success we need also to reset AZ
which defined in the provider_client. So any logic will get into correct AZ.