usb: Add usb-devices check for usb_drive
For now the lsscsi has been used with no check if the usb subsystem
detects the usb device. With this patch there is an optional usb serial
check to see if usb-devices detects whitelisted and thus expected
usb device being recognized and initialized.

The patch is proposed based on the example of the test run which
indicates that 'no usb storage device connected' using lsscsi, however
the logs clearly show that the usb device was detected with the
following logs:
[ T215] usb 1-11: Product: SanDisk 3.2Gen1
[ T215] usb 1-11: Manufacturer: USB
[ T215] usb 1-11: SerialNumber: 040170f1881558....
[ T215] usb 1-11: can't set config #1, error -32

Proposed patch helps investigating such situations