Fix inconsistent check for plain text passwords
The unit test for "Passwords in the data folder" was dependent on the system's `grep` version,
causing it to fail on some environments, such as openSUSE Tumbleweed with grep 3.11.
To ensure consistent behavior between local and CI/CD environments,
change the test on the grep output only considering stdout and not the exit code.