Adapt to new behaviour of hpc-testing
Previously, hpc-testing for InfiniBand had default values for the IP
address configuration of IPoIB interfaces. This default went away, so
we need to specify these. They are variables to allow overriding in
case we need it, but they should only be used internally for the test
suite and not be of relevance outside of this - just make it show the
old behaviour again.