containers: Fix container_images test
FIPS tests are still failing so a different approach is needed.
We can't rely on curl --retry because we can't use the more recent
set of --retry-xxx options as they're not available on older products,
and getting the curl version would be a mess.

Also clean up other parts of the code.