kselftests: Support both -t and -c option
Ideally the tests allow to exclude some known to fail tests, however
run_selftests.sh, a native kernel selftest run script, does not provide
such option. The way to handle that is to run only specific suites
with the -t option, so any other downstream test executor can whitelist
the tests to run from a given test group. This patch allows to use that
so faulty tests might be not scheduled for the given codestream