Accept Promoted/Unpromoted in drbd_passive test module
Recent versions of crmsh have replaced the terms Master/Slave with
Promoted/Unpromoted. The `ha/drbd_passive` test module fails as it tries
to confirm the `ms_drbd_passive` resource is started and running in the
first node by looking for `Master` in the output of `crm resource
status`. This commit makes it so both `Master` and `Promoted` are
accepted.

Also included in the commit is an added command in
`ha/iscsi_client_setup` to restart `systemd-udevd` before iSCSI
configuration in SLES 16, as now udev comes configured with
`ProtectHostname=yes` and without restarting udev or the node, it's
possible that the hostname seen by udev and the one configured in the
system do not match, leading to problems later in the test.

Additionally, also included in the commit is a call to
`corosync-cmapctl` added with `script_retry` in
`hacluster::collect_sbd_delay_parameters` as the function requires the
cmap API to be ready. Retrying the command before attemtpting to collect
the SBD params allows us to check whether the cmap API did start.

Related Ticket: https://bugzilla.suse.com/show_bug.cgi?id=1247534