lib: Fix script_retry dies on timeout
Adds a safeguard for the timeout command to kill the underlying command
if stuck.

Adding the "kill after 5 seconds" option to timeout should be relatively
safe, because without this the test would fail with a timeout, and this
gives the system under test another retry to do run a command. If the
system is borked, it will very likely fail on those subsequent tries.