Mark 'empty repos after registration' with soft-failure
This commit introduces a check for the zypper exit code in a publiccloud
related zypper call so that we can check for the specific exit code and
record a reference to a known bug on detection. This commit also
introduces the dependency on Feature::Compat::Try which is unproblematic
as the same dependency is already mandatory for os-autoinst anyway.