New installation method for .tar.xz images
As stated by developers, Microsoft is deprecating the .appx format in favor of .tar.xz. These new images are already being uploaded to the Microsoft Store and are what we've been referring to as "modern" images.

The aim of this ticket is to formalize a new, tested procedure for extracting and installing future .tar.xz images when needed. This method has been verified locally and is a straightforward, three-step process.