sudo: Fix search for password prompt properly
Instead of a racy and wasteful "sleep 1" as introduced in 67d2102e2 one
should expect the correct password prompt. Previously the expect script
only looked for the string password which can also appear in the
preample text of an interactive sudo call. This was the reason for the
error that appeared in https://progress.opensuse.org/issues/186353 where
the text "For security reasons, the password you ..." was erroneously
detected as the password prompt and expect immediately typed the
password as is also visible on the screen of the affected test. Instead
this commit removes the "sleep 1" call and uses the unique string
"password for" to look for.

Related progress issue: https://progress.opensuse.org/issues/186474