publiccloud/run_ltp: Remove obsolete --framework ltp
kirk --framework option was removed in [1] (released in kirk 2.1), now
kirk supports only LTP. It might be deleted in the future therefore safe
to remove it.

In fact it was never needed for LTP, because LTP was the default
framework, therefore it will work on older kirk (which supported more
testsuites via --framework option).

[1] https://github.com/linux-test-project/kirk/commit/515a5053170a04d3b6f9db0be7217f66fb6cfcb8

Signed-off-by: Petr Vorel <pvorel@suse.cz>