Limit timeout for supportconfig
If supportconfig fails, it internally restarts. This means that if it
fails over and over, it will just run until it hits a timeout (2h).
Since there is effectively no progress, this timeout is way too high.

If after 10 minutes the supportconfig is not yet done, it's safe to
assume it's stuck and should be cancelled.