Update the test data of repos for medium installation testsuites
It is quite close to sle16 GMC now, it is reasonable to remove the
'Beta' string from the repos' name. So update the test data of repos for
media installation. BTW, there isn't 'Beta' string in the installation
cases in Online flavor, so no need to update datas for Online
testsuites.