bats: Also generate JUnit report
In addition to TAP files, also generate a JUnit report so we can
eventually migrate from TAP to JUnit to mitigate for the lack
of fine-grained soft-failures in TAP.  We'll use TAP for human
consumption and JUnit for openQA processing.

For now let's gather these XML files that will help create a list
of tests to ignore or soft-fail, given that we use a very broad
mechanism that ignores the failures in whole files in the form of
the BATS_IGNORE variables.

- Related ticket: https://progress.opensuse.org/issues/181616