Kselftests: Introduce KSELFTEST_RUNNER
It might be useful to provide a custom runner instead of always relying on
the standard `run_kselftest.sh` script due to limitations on how it deals
with passing through arguments. For example, when investigating issues
regarding a specific test such as "bpf:test_progs" which is a full binary
on its own, it might be useful to pass the `-v` option to it which is not
possible using the standard runner.