Mark bug to the autoyast profile of sap 15SP7 installation
We added two autoyast profiles in pr#23209, they are almost the
same except the <default_target> setting. The reason is there's
a bug bsc#1235701 on SLE15SP7 instalaltion with gnome desktop on
ppc64le, which will not be fixed on SLE15SP7.
Mark the bug number to the profile name as a reminder.