Add test module to check iSCSI initiator randomness
This commit adds a new schedule and test module to verify that the
default iSCSI initiator configured in the SUT by agama is different
after each installation.

This is in response to
https://bugzilla.suse.com/show_bug.cgi?id=1246280, whose root cause was
determined to be the installer copying in `/etc/iscsi/initiatorname.iscsi`
the same value after all installations, and thus violating the
uniqueness required for iSCSI initiators.

The new schedule performs 2 installations back to back on the same SUT.
Right after each installation, the new module `ha/check_iscsi_initiator`
is scheduled to verify the value of the initiator. On the first run, it
will simply store the value in a setting; in the second run it will
compare the stored value against the one from the second installation
and either succeed of fail accordingly.

Related Ticket: https://jira.suse.com/browse/TEAM-10542