Upload ipaddr2 logs on failure
This change introduces a new subroutine `ipaddr2_logs_collect` that
collects and uploads logs when a job fails.

The following logs are collected:
- journal
- crm report
- y2_logs
- supportconfig

The `post_fail_hook` in all the `ipaddr2` tests is modified to call
this new subroutine.
The implementation also include an internal function similar to
qesap_cluster_log_cmd, to eventaully be able to merge then in a
dedicated log related library in the future.