Test improvements for FIPS env mode
Skip shutdown for security ipmi tests. Subsequent tests will re-use
the machine, so there is no real need to shut it down.

Add --may-exist flag to ovs-vsctl command in the openvswitch test.

Double check the FIPS mode and do not rely on openQA env vars only.

This is also an attempt to fix poo#183539.

Signed-off-by: Ferenc Szekely <feri.szekely@suse.com>