Dont expected grub2 screen in vmware
sle-micro relies on precise handling of grub2 screen in openQA.
This is a problem in case of VMware as we have a race condition, due to
the fact that the related backed reconnects vnc after a certain stalled period.

The problem occurs when openQA reconnects VNC, the grub2 screen is
available, but the needles match early and key strokes might be missed.

- ticket: [[MinimalVM] VMware reboots are often causing needle mismatches](https://progress.opensuse.org/issues/188997)