Extend get_gcp_guest_os_features to 16.0 for SAP
Extend image name regexp to also consider sles4sap GCE images.
Provide a set of os_features for the 16.0 SAP image, that at the moment
is simply a copy of the SLES-16.0 one.
Purpose of this change is to provide right value for the --guest-os-features
argument in the `gcloud compute images create` command.