Drop timeout => 0
In such a case script_run will just blindly hit enter and hope for the best.
This is not acceptable because when things goes wrong this ruins further flow completely.
Command with which hanging is not correctly finished so we start to type next command
while being within of execution of previous