bats: umoci: fix skopeo copy invocation
commit c190d0e63400 ("bats: Add umoci upstream tests") incorrectly
called skopeo copy with no tag specified for the OCI image target. This
causes skopeo to create an OCI image with no top-level tag and thus
umoci could not find the "latest" image to use for the tests:

  % skopeo copy \
    docker://registry.opensuse.org/opensuse/tumbleweed:latest \
    oci:/var/tmp/image
  ...
  % jq . /var/tmp/image/index.json
  {
    "schemaVersion": 2,
    "manifests": [
      {
        "mediaType": "application/vnd.oci.image.manifest.v1+json",
        "digest": "sha256:53d3c0ebd12dcde196789266fca96dc673ee15a093dfaf758d1b2012a6f5d075",
        "size": 407
      }
    ]
  }

The solution is to simply specify a target tag (while we're at it, add
SOURCE_TAG to the bats invocation to make it clear what tag we are
using for the tests).

  % skopeo copy \
    docker://registry.opensuse.org/opensuse/tumbleweed:latest \
    oci:/var/tmp/image:latest
  % jq . /var/tmp/image/index.json
  {
    "schemaVersion": 2,
    "manifests": [
      {
        "mediaType": "application/vnd.oci.image.manifest.v1+json",
        "digest": "sha256:53d3c0ebd12dcde196789266fca96dc673ee15a093dfaf758d1b2012a6f5d075",
        "size": 407,
        "annotations": {
          "org.opencontainers.image.ref.name": "latest"
        }
      }
    ]
  }

The vast majority (if not all) of the test failures with umoci were due
to this mistake:

  # umoci unpack --image /tmp/umoci-integration/.../image:latest ... (status=1)
  #    ⨯ tag is not found: latest

Fixes: c190d0e63400 ("bats: Add umoci upstream tests")
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>