Remove mellanox_config from schedules (#23817)
The module is actually only needed when the ConnectX-5 cards were
configured into Ethernet mode.
However, we don't run any tests that do this, so we can save around 5
minutes from avoiding a reboot we don't need. If we want to use Ethernet
on these machines, we can use the other Mellanox cards that are plain
Ethernet.

Co-authored-by: Michael Moese <mmoese@suse.com>