Upload more HANA installation logs in CSP tests
On HanaSR tests in the cloud, current code collects `hdblcm.log` and
`hdbinst.log` from the installation and uploads them to the job; however
this is some times not enough to do a full debug and identify root
causes, as both logs usually point to other logs and trace files. For
example is very common for `hdblcm.log` to reference errors in
`hdbinst_configure.log`, and even that file can reference other trace
files.

This commit adds in the log gathering function the generation of a
tarball of the whole installation log directory for it to be uploaded as
part of the results.