Drop LTP_SUITE_TIMEOUT in favor of LTP_TIMEOUT
Currently we have two timeouts one which defines how long openQA waiting for kirk to finish
and another is how long kirk itself waiting for test to finish. Obviously first should always
be bigger than the second. To simplify this we will use single variable for both with unconditional
extension in first case