No assert in all delete aws_cli functions (#23996)
Change assert_script_run to script_run and return exit code in all
aws_cli function usually involved in infrastructure cleanup. Leave to
the caller the responsability to handle the failure, allowing the
cleanup procedure to be more smart and delete more resources.
Crash test cleanup use it to delete as much AWS resources as possible
and not stopping the cleanup procvess at the first failure.