Proposal to use iSCSI LUNs created on demand on non qemu jobs
This commit introduces a special value for the ISCSI_LUN_INDEX setting
(`ondemand`), so that the `ha/iscsi_client_setup` test module, requests a
number of iSCSI LUNs to be created in the iSCSI server on demand when
`ISCSI_LUN_INDEX=ondemand`.

Implementation consists of creating a specially named directory in the
`ondemand` subdir located in the share referenced by NFS_SUPPORT_SHARE.
The directory name is the combination of the openQA instance where job
was scheduled, the job id and the number of LUNs.

A process in the iSCSI server will then read this directory, and setup
the new target and LUNs; meanwhile the `ha/iscsi_client_setup` test module
will wait for the target to appear by checking with `iscsiadm`.

Related Ticket: https://jira.suse.com/browse/TEAM-10796