nfs_stress_ng: Mark the failure as fatal
If nfs_stress_ng it has an effect on the following modules, as it causes
the next module to attempt to rollback. Seems like inevitbale behaviour
in the multimachine tests, thus the module should be marked as fatal.
Rollback in the multimachine scenarios are problematic for multiple
reasons, for instance all network states or nfs mounts among others