Run sestatus in serial terminal in sles4sap::pre_run_hook always
This fixes a race condition in `sles4sap::pre_run_hook()` when the code
runs on a SUT where the `root-console` is currently selected, but the
SUT has been rebooted, which causes the `script_output('sestatus')`
command to fail as it's not written in a shell session. Solution was to
always run the `sestatus` command from the serial terminal, and return
to the previous console if there was a console selected. It has a bit of
an overhead as now `pre_run_hook` will always switch the console to the
serial terminal (unless it was already selected), but other solutions
would've required changes in more places.